bookitrace/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pandas==1.3.3
scikit-surprise==1.1.1